requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
pytz==2024.2
gunicorn==21.2.0
openpyxl==3.1.2
flask-mail==0.9.1
pyjwt==2.8.0
qrcode==7.4.2
pyotp==2.9.0